[FIX] Fix text input autofocus not working on Firefox
